ml/mcp/requirements.txt
# ATLAS MCP Agent — Self-contained dependencies
# Install: pip install -r requirements.txt

# MCP protocol
mcp[cli]>=1.0.0

# LLM backends (modular — use either or both)
openai>=1.30.0           # OpenAI / any OpenAI-compatible API
llama-cpp-python>=0.2.0  # Local models (optional, for offline)

# Browser automation
playwright>=1.40.0

# Utilities
httpx>=0.27.0
pydantic>=2.0.0
rich>=13.0.0             # Pretty console output
python-dotenv>=1.0.0     # .env file loading
